removed login activity.
