+++ app/src/main/res/drawable/button_welcome.xml
@@ -1,0 +1,11 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@android:color/white"/>
+    <corners
+        android:bottomRightRadius="70dp"
+        android:bottomLeftRadius="70dp"
+        android:topLeftRadius="70dp"
+        android:topRightRadius="70dp" />
+    <stroke
+        android:color="@color/colorGrey"
+        android:width="4dp" />
+</shape>
